+++ src/index.css
@@ -1,4 +1,4 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
 
-@import './components/styles/main-home.css';
+@import './components/Home/styles/main-home.css';
